fix(TFD-4606): Separate Beam and TCK/JSON processors.
