+++ pom.xml
@@ -47,7 +47,7 @@
     <surefire.version>2.21.0</surefire.version>
     <jacoco.version>0.8.1</jacoco.version>
 
-    <!-- set to true if you use native beam programming model -->
+    <!-- set to false if you use native beam programming model -->
     <validation.model>true</validation.model>
   </properties>
 
@@ -56,7 +56,8 @@
     <module>fileio</module>
     <module>bigquery</module>
     <module>elasticsearch</module>
-    <module>processing</module>
+    <module>processing-beam</module>
+    <module>processing-tck</module>
     <module>jdbc</module>
     <module>salesforce</module>
   </modules>
