+++ processing-beam/src/main/java/org/talend/components/processing/replicate/ReplicateConfiguration.java
