+++ processing-tck/src/main/java/org/talend/components/processing/fieldselector/FieldSelector.java
