+++ processing-tck/src/main/java/org/talend/components/processing/fieldselector/FieldSelectorConfiguration.java
