+++ processing-tck/src/main/java/org/talend/components/processing/filterrow/FilterRowConfiguration.java
