+++ processing-tck/src/main/java/org/talend/components/processing/limit/LimitConfiguration.java
