+++ processing-tck/src/main/java/org/talend/components/processing/normalize/Normalize.java
