+++ processing-tck/src/main/java/org/talend/components/processing/package-info.java
@@ -1,0 +1,6 @@
+@Icon(Icon.IconType.FILTER)
+@Components(family = "Processing")
+package org.talend.components.processing;
+
+import org.talend.sdk.component.api.component.Components;
+import org.talend.sdk.component.api.component.Icon;
