+++ processing-tck/src/main/resources/org/talend/components/processing/normalize/Messages.properties
