+++ processing-tck/src/test/java/org/talend/components/processing/fieldselector/FieldSelectorTest.java
