+++ processing-tck/src/test/java/org/talend/components/processing/normalize/NormalizeTest.java
