--- processing/pom.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="
-          http://maven.apache.org/POM/4.0.0
-          http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.talend.components</groupId>
-    <artifactId>connectors-se</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>processing</artifactId>
-  <name>Components :: Processing</name>
-
-  <properties>
-    <!-- native beam -->
-    <validation.model>false</validation.model>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.apache.beam</groupId>
-      <artifactId>beam-sdks-java-core</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.beam</groupId>
-      <artifactId>beam-runners-direct-java</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.talend.sdk.component</groupId>
-      <artifactId>component-runtime-beam-junit</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.talend.daikon</groupId>
-      <artifactId>daikon</artifactId>
-      <version>0.23.0</version>
-      <classifier>tests</classifier>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
-      <version>2.21.0</version>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.talend.components</groupId>
-      <artifactId>components-adapter-beam</artifactId>
-      <version>${tcomp0.version}</version>
-      <exclusions>
-        <exclusion>
-          <groupId>org.apache.beam</groupId>
-          <artifactId>beam-sdks-java-extensions-google-cloud-platform-core</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>org.apache.beam</groupId>
-          <artifactId>beam-sdks-java-core</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>org.apache.beam</groupId>
-          <artifactId>beam-runners-direct-java</artifactId>
-        </exclusion>
-        <!--<exclusion>-->
-          <!--<groupId>org.talend.daikon</groupId>-->
-          <!--<artifactId>daikon</artifactId>-->
-        <!--</exclusion>-->
-      </exclusions>
-    </dependency>
-    <!-- guava for components-adapter-beam -->
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <version>23.0</version>
-    </dependency>
-  </dependencies>
-
-</project>
